Finally got damn info from firebase
